feat: Refactor credential handling and extraction logic for status attestation
