+++ src/credential/status/02-status-attestation.ts
@@ -4,11 +4,7 @@ import {
   type Out,
 } from "../../utils/misc";
 import type { EvaluateIssuerTrust, ObtainCredential } from "../issuance";
-import {
-  type CryptoContext,
-  SignJWT,
-  thumbprint,
-} from "@pagopa/io-react-native-jwt";
+import { type CryptoContext, SignJWT } from "@pagopa/io-react-native-jwt";
 import { v4 as uuidv4 } from "uuid";
 import { StatusAttestationResponse } from "./types";
 import {
@@ -18,8 +14,7 @@ import {
   UnexpectedStatusCodeError,
 } from "../../utils/errors";
 import { Logger, LogLevel } from "../../utils/logging";
-import { CBOR } from "@pagopa/io-react-native-cbor";
-import { decode } from "../../sd-jwt";
+import { extractJwkFromCredential } from "../../utils/credentials";
 
 export type StatusAttestation = (
   issuerConf: Out<EvaluateIssuerTrust>["issuerConf"],
@@ -108,21 +103,4 @@ const handleStatusAttestationError = (e: unknown) => {
       message: `Unable to obtain the status attestation for the given credential`,
     })
     .buildFrom(e);
-};
-
-const extractJwkFromCredential = async (credential: string) => {
-  if (credential.includes("~")) {
-    const parsed = decode(credential);
-    const jwk = parsed.sdJwt.payload.cnf.jwk;
-    if (jwk) {
-      return { ...jwk, kid: await thumbprint(jwk) };
-    }
-  }
-
-  // 2. CBOR case
-  const decoded = await CBOR.decode(credential);
-  const jwk = decoded?.credentialSubject?.cnf?.jwk;
-  if (jwk) {
-    return { ...jwk, kid: await thumbprint(jwk) };
-  }
 };
